[ALPS08058990] camsys: fix coverity
[Description]
1. fix coverity

[Test]
Build pass and self-sanity pass

MTK-Commit-Id: b3b8bd0738648f758d1055b8e6bb5168dad7bedd

CR-Id: ALPS08058990
Feature: [Android Default] Camera Application Basic Functions

+++ camsys/isp8/cam/mtk_cam-job.c
@@ -4718,7 +4718,7 @@ static void check_sen_expo_change(struct mtk_cam_job *job)
 
 	if (likely(qof_get_mtcmos_margin() > MARGIN_TO_VOTE_QOF_US)) {
 		voter_on_margin_ns =
-			(qof_get_mtcmos_margin() - MARGIN_TO_VOTE_QOF_US) * -1000;
+			(s64)(qof_get_mtcmos_margin() - MARGIN_TO_VOTE_QOF_US) * -1000;
 	} else {
 		WARN_ON(1);
 		return;
